Reorganizaçao das pastas de componentes v1.0.0
